refactor: refactor graph
